packages/doeff-seedream/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doeff-seedream"
version = "0.1.0"
description = "ByteDance Seedream integration for doeff with image generation helpers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CyberAgent AI Lab"},
]
keywords = ["seedream", "bytedance", "effects", "image-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "doeff>=0.1.0",
    "httpx>=0.27",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/CyberAgentAILab/doeff"
Repository = "https://github.com/CyberAgentAILab/doeff"
Issues = "https://github.com/CyberAgentAILab/doeff/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/doeff_seedream"]

[tool.uv.sources]
doeff = {workspace = true}

[dependency-groups]
dev = []
